--- Prototype Week1/Assets/MenuInicial.cs
@@ -1,17 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class MenuInicial : MonoBehaviour
-{
-    public void Jugar()
-    {
-        SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
-    }
-    public void Salir()
-    {
-        Debug.Log("Salir..");
-        Application.Quit();
-    }
-}
