--- Prototype Week1/Assets/MenuInicial.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3e12a025828b9a3438e106e9e32ab8b7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
