+++ Prototype Week1/Assets/Scripts/MenuInicial.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class MenuInicial : MonoBehaviour
+{
+    public void Jugar()
+    {
+        SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
+    }
+    public void Salir()
+    {
+        Debug.Log("Salir..");
+        Application.Quit();
+    }
+}
