+++ Prototype Week1/Assets/Scripts/MenuOpciones.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.Audio;
+
+public class MenuOpciones : MonoBehaviour
+
+{
+    [SerializeField] private AudioMixer audioMixer;
+
+    public void PantallaCompleta(bool pantallaCompleta)
+    {
+        Screen.fullScreen = pantallaCompleta;
+    }
+
+    public void CambiarVolumen(float volumen)
+    {
+        audioMixer.SetFloat("Volumen", volumen);
+    }
+}
