+++ Prototype Week1/Assets/Scripts/MenuOpciones.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73a9b39475957d34fa1ea92beaf21980
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
